+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 [Unreleased]: https://github.com/chaostoolkit-incubator/chaostoolkit-aws/compare/0.21.2...HEAD
 
+### Changed
+- Added new EC2 probe to check whether minimum number of instances are running
+
 ## [0.21.2][] - 2022-07-12
 [0.21.2]: https://github.com/chaostoolkit-incubator/chaostoolkit-aws/compare/0.21.1...0.21.2